+++ src/middleware/index.ts
@@ -1,3 +1,3 @@
-export * from "./auth-validation";
+export * from "./auth/validate-body";
 export * from "./rate-limit";
-export * from "./auth-authorization";
+export * from "./auth/auth-authorization";
